backend/src/requirements.txt
aioredis==1.3.1
amqp==5.1.1
appnope==0.1.3
APScheduler==3.9.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.6.0
awsebcli==3.20.2
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
blessed==1.19.1
boto3==1.18.0
botocore==1.21.65
CacheControl==0.12.11
cached-property==1.5.2
cachetools==5.2.0
celery==5.2.7
cement==2.8.2
certifi==2022.6.15
cffi==1.15.0
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.13
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-ebhealthcheck==2.0.1
django-extensions==3.1.5
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.13.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.4
fastjsonschema==2.15.3
filelock==3.7.1
firebase-admin==5.2.0
future==0.16.0
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.20.4
huggingface-hub==0.8.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.4
incremental==21.3.0
ipykernel==6.15.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kombu==5.2.4
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.4
names==0.3.0
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.6.2
requests==2.26.0
rsa==4.8
s3transfer==0.5.2
scipy==1.7.3
semantic-version==2.8.5
Send2Trash==1.8.0
sendgrid==6.9.7
service-identity==21.1.0
six==1.14.0
slack-bolt==1.15.1
slack-sdk==3.19.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==4.1.0
supervisor==4.2.4
termcolor==1.1.0
terminado==0.15.0
texttable==1.6.4
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
torch==1.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
watchtower==3.0.0
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.59.0
widgetsnbextension==3.6.1
zipp==3.8.0
zope.interface==5.4.0
